pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["woojin-an <dkswoojin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {version = "^0.115.0", extras = ["standard"]}
uvicorn = "^0.32.0"
gunicorn = "^20.1.0"
pydantic-settings = "^2.6.1"
tortoise-orm = {version = "^0.21.7", extras = ["asyncmy"]}
aerich = "^0.7.2"
cryptography = "^43.0.3"
greenlet = "^3.1.1"
python-multipart = "^0.0.17"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
redis = "^5.2.0"
boto3 = "^1.35.57"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
coverage = "^7.6.4"
pytest = "^8.3.3"


[tool.aerich]
tortoise_orm = "src.core.database.db_settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.12"
strict = true
